requirements.txt
numpy~=1.19.3
pandas~=1.4.4
scikit_learn~=1.2.2
torch~=1.11.0
tqdm~=4.65.0
pyyaml
nni
xgboost